[BUGFIX] Tx_Phpunit_Service_DatabaseTest::selectReturnsRessource Test fails on 6.1+
A select on TYPO3_DB does not return a resource anymore due to the change to mysqli.
It is a mysqli_result instance now.

The test should therefore only be executed when running on a pre-6.1 installation.
For TYPO3 CMS 6.1 an up, the test needs to check for the mysqli_result instance.

Resolves: #49841

Change-Id: I0473760803233d12fc7b88cc832d8a36f536aaec
Reviewed-on: https://review.typo3.org/22425
Reviewed-by: Nicole Cordes
Tested-by: Nicole Cordes
Reviewed-by: Oliver Klee
Tested-by: Oliver Klee