[CLEANUP] Change the test case class comments to "Test case."
Mentioning the name of the tested class in the test case class PHPDoc is
redundant because the test case class name already contains the name of
the tested class.

In addition, removing the class name from the test case PHPDoc makes our
work easier when we want to rename a class.

Resolves: #50211

Change-Id: I92338193d9a668d2c13d35338848ac748717ae63
Reviewed-on: https://review.typo3.org/22435
Reviewed-by: Nicole Cordes
Tested-by: Nicole Cordes