[BUGFIX] Ignore missing extensions
If an extension directory is removed, TYPO3 keeps the extension in
the list of loaded extensions, but any call related to the extension,
e.g. t3lib_extMgm::extPath, will result in an exception.

Ignore any exceptions during test finding and treat missing or broken
extensions as extension without tests.

Fixes: #40809
Change-Id: Ica27f638f1e0ebc86d4b773b2e0fc1d2dca7f7c7
Reviewed-on: https://review.typo3.org/31195
Reviewed-by: Oliver Klee
Tested-by: Oliver Klee