[TASK] Add compat to TYPO3 CMS 7 / Drop t3lib-usage
To achieve compatibility with TYPO3 CMS 7 (current master) drop
the usage of t3lib_*/tslib_* since no compat-class-maps for those
classes exist anymore.

This also means that we raise the minimal version to 6.2.4.

We also can drop the special treatment for the Core tests as those
now are part of normal system extensions.

Change-Id: I3ac2959554525950564b7af88bc48dbc320c9da9
Resolves: #62788
Reviewed-on: http://review.typo3.org/33908
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Tested-by: Oliver Klee <typo3-coding@oliverklee.de>
Reviewed-by: Stefan Neufeind <typo3.neufeind@speedpartner.de>
Tested-by: Stefan Neufeind <typo3.neufeind@speedpartner.de>